Remove db_migration0.sql, update db_migration.sql
The db_migration0.sql file has been deleted and the db_migration.sql file has been updated. The db_migration.sql file now includes changes to the schema of the celestial bodies and person tables, as well as the creation of an index for the username column in the person table.